requirements.txt
Flask==1.1.2
Pillow>=3.0.0
waitress==2.0.0
